% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_dicc.R
\name{pjp_dicc}
\alias{pjp_dicc}
\title{returns a df with el diccionario de un df}
\usage{
pjp_dicc(df, truncate = TRUE, nn = 100)
}
\arguments{
\item{df}{(a data frame)}

\item{truncate}{(default to TRUE)}

\item{nn}{(if you truncate, this is the maximun character that you will see. The default = 100)}
}
\value{
Un nuevo \code{df} con los valores unicos del df original
}
\description{
pjp_f_dicc() returns a df con el diccionario del input df
}
