% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_estadisticos_basicos.R
\name{pjp_estadisticos_basicos}
\alias{pjp_estadisticos_basicos}
\title{Crea un resumen de un data.frame.}
\usage{
pjp_estadisticos_basicos(data, print = FALSE)
}
\arguments{
\item{data}{(a data frame)}

\item{print}{(default to FALSE)}
}
\description{
Para cada variable del df calcula: número y porcentaje de valores cero (q_zeros and p_zeros respectevly). Same metrics for NA values (q_NA and p_na). Last two columns indicates data type and quantity of unique values.
This function uses a df as inputs and returns a dataframe with the names and some chracteristics of the variables (colums) of the input dataframe. By the way, this function comes from the funModeling package made by Pablo Casas \href{mailto:pabloc@datascienceheroes.com}{pabloc@datascienceheroes.com}, gracias Pablo!! Solo le he añadido la opcion por defecto print=FALSE y modificado alguna cosa menor. Lo dejo tal cual para no tener dependencias
This function print and return the results.
}
\examples{
pjp_estadisticos_basicos(cars)
}
