% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_not_in.R
\name{pjp_not_in}
\alias{pjp_not_in}
\title{Mi versión de not_in}
\usage{
pjp_not_in(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
Un nuevo vector con los elementos de x que no están en y
}
\description{
Te dice los valores de x que no están en y
}
\examples{
pjp_not_in(c(1,2,3), c(2,3))
}
