% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_open_urls.R
\name{pjp_open_urls}
\alias{pjp_open_urls}
\title{Abre un grupo de links en el browser}
\usage{
pjp_open_urls(urls = "https://www.google.com")
}
\arguments{
\item{urls}{vector con url's a  abrir}
}
\description{
Proporcionas un vector con url's y las abre en el navegador
}
\examples{
pjp_open_urls("https://perezp44.github.io/intro-ds-21-22-web/")
}
