% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_unique_values.R
\name{pjp_unique_values}
\alias{pjp_unique_values}
\title{Retorna un data.farme con los valores únicos de cada columna}
\usage{
pjp_unique_values(df, truncate = FALSE, nn = 100)
}
\arguments{
\item{df}{(a data frame)}

\item{truncate}{(default to FALSE)}

\item{nn}{(if you truncate, this is the maximun character that you will see. The default = 100)}
}
\value{
Un nuevo \code{df} con los valores unicos del df original
}
\description{
En realidad pjp_unique_values() retorna a df con 3 columnas:
\enumerate{
\item el nombre de las variables, 2) El nº de valores únicos de cada columna, y 3) los valores únicos de las variables
}
}
\examples{
pjp_unique_values(iris)
}
