% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_theme_maps.R
\name{theme_pjp_maps}
\alias{theme_pjp_maps}
\title{Un theme apropiado para mapas}
\usage{
theme_pjp_maps()
}
\value{

}
\description{
theme_pjp_maps() simplemente crea un theme apropiado para mapas
}
\details{
\link{ggplot2 theme} apropiado para mapas
}
